backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
annoy==1.17.3
chromaprint==0.7
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2